Compilation errors to correct

+++ minilab_manager/include/minilab_manager/interface.hpp
@@ -39,7 +39,7 @@ private:
   bool onManualRequest(Empty::Request &req, Empty::Response &res);
   bool isMapOk();
   bool isTrajectoryOk();
-  void updateState(uint8 state);
+  void updateState(uint state);
 };
 
 }
